feat [IN] transaction

+++ service/merchant_service.go
@@ -10,11 +10,11 @@ import (
 )
 
 type MerchantService interface {
-	Create(ctx context.Context, request web.MerchantCreateRequest, image interface{}) web.MerchantCreateRequest
+	Create(ctx context.Context, request web.MerchantCreateRequest) web.MerchantCreateRequest
 	FindById(ctx context.Context, merchantId string) web.MerchantResponse
 	FindManageOrderById(ctx context.Context, merchantId string) web.ManageOrderResponse
 	Update(ctx context.Context, request web.MerchantUpdateRequest) web.MerchantUpdateRequest
-	UpdateMainImage(ctx context.Context, request web.MerchantUpdateImageRequest, image interface{}) web.MerchantUpdateImageRequest
+	UpdateMainImage(ctx context.Context, request web.MerchantUpdateImageRequest) web.MerchantUpdateImageRequest
 	Delete(ctx context.Context, merchantId string)
 }
 
@@ -32,8 +32,9 @@ func NewMerchantService(merchantRepository repository.MerchantRepository, cloudi
 	}
 }
 
-func (service *merchantServiceImpl) Create(ctx context.Context, request web.MerchantCreateRequest, image interface{}) web.MerchantCreateRequest {
-	url, err := service.CloudinaryRepository.UploadImage(ctx, request.MainImage.FileName, image)
+func (service *merchantServiceImpl) Create(ctx context.Context, request web.MerchantCreateRequest) web.MerchantCreateRequest {
+
+	url, err := service.CloudinaryRepository.UploadImage(ctx, request.MainImage.FileName, request.MainImage.URL)
 	helper.PanicIfError(err)
 
 	_, err = service.MerchantRepository.Create(ctx, domain.Merchant{
@@ -42,6 +43,7 @@ func (service *merchantServiceImpl) Create(ctx context.Context, request web.Merc
 		Email:     request.Email,
 		Password:  helper.HashPassword(request.Password),
 		Name:      request.Name,
+		Slug:      request.Slug,
 		Phone:     request.Phone,
 		MainImage: &domain.Image{
 			Id:       primitive.NewObjectID(),
@@ -54,12 +56,15 @@ func (service *merchantServiceImpl) Create(ctx context.Context, request web.Merc
 			Province:   request.Address.Province,
 			Country:    request.Address.Country,
 			PostalCode: request.Address.PostalCode,
-			Latitude:   request.Address.Latitude,
-			Longitude:  request.Address.Longitude,
 		},
 	})
-	helper.PanicIfError(err)
+	if err != nil {
+		errUpload := service.CloudinaryRepository.DeleteImage(ctx, request.MainImage.FileName)
+		helper.PanicIfError(errUpload)
+		panic(err.Error())
+	}
 
+	request.MainImage.URL = url
 	return request
 }
 
@@ -74,7 +79,9 @@ func (service *merchantServiceImpl) FindById(ctx context.Context, merchantId str
 		UpdatedAt: res.UpdatedAt,
 		Email:     res.Email,
 		Name:      res.Name,
+		Slug:      res.Slug,
 		Phone:     res.Phone,
+		Balance:   res.Balance,
 		MainImage: &web.ImageResponse{
 			Id:       res.MainImage.Id.Hex(),
 			FileName: res.MainImage.FileName,
@@ -86,8 +93,6 @@ func (service *merchantServiceImpl) FindById(ctx context.Context, merchantId str
 			Province:   res.Address.Province,
 			Country:    res.Address.Country,
 			PostalCode: res.Address.PostalCode,
-			Latitude:   res.Address.Latitude,
-			Longitude:  res.Address.Longitude,
 		},
 	}
 }
@@ -96,31 +101,40 @@ func (service *merchantServiceImpl) FindManageOrderById(ctx context.Context, mer
 	merchant, err := service.MerchantRepository.FindById(ctx, merchantId)
 	helper.PanicIfError(err)
 
-	var productsRespone []*web.ManageOrderProductResponse
+	var productsResponse []*web.ManageOrderProductResponse
 	for _, v := range merchant.Orders {
 		product, err := service.ProductRepository.FindById(ctx, v.ProductId)
 		helper.PanicIfError(err)
 
-		productsRespone = append(productsRespone, &web.ManageOrderProductResponse{
+		productsResponse = append(productsResponse, &web.ManageOrderProductResponse{
 			Id:          v.Id.Hex(),
 			CreatedAt:   v.CreatedAt,
 			UpdatedAt:   v.UpdatedAt,
 			ProductId:   product.Id.Hex(),
 			Name:        product.Name,
+			Slug:        product.Slug,
 			Description: product.Description,
 			Price:       v.Price,
+			Quantity:    v.Quantity,
+			SubTotal:    v.Price * v.Quantity,
 			MainImage: &web.ImageResponse{
 				Id:       product.MainImage.Id.Hex(),
 				FileName: product.MainImage.FileName,
 				URL:      product.MainImage.URL,
 			},
-			Quantity: v.Quantity,
+			Address: &web.AddressResponse{
+				Address:    v.Address.Address,
+				City:       v.Address.City,
+				Province:   v.Address.Province,
+				Country:    v.Address.Country,
+				PostalCode: v.Address.PostalCode,
+			},
 		})
 	}
 
 	return web.ManageOrderResponse{
 		MerchantId: merchant.Id.Hex(),
-		Products:   productsRespone,
+		Products:   productsResponse,
 	}
 }
 
@@ -132,6 +146,7 @@ func (service *merchantServiceImpl) Update(ctx context.Context, request web.Merc
 		Id:        merchant.Id,
 		UpdatedAt: request.UpdatedAt,
 		Name:      request.Name,
+		Slug:      merchant.Slug,
 		Phone:     request.Phone,
 		Address: &domain.Address{
 			Address:    request.Address.Address,
@@ -139,24 +154,23 @@ func (service *merchantServiceImpl) Update(ctx context.Context, request web.Merc
 			Province:   request.Address.Province,
 			Country:    request.Address.Country,
 			PostalCode: request.Address.PostalCode,
-			Latitude:   request.Address.Latitude,
-			Longitude:  request.Address.Longitude,
 		},
 	})
 	helper.PanicIfError(err)
 	return request
 }
 
-func (service *merchantServiceImpl) UpdateMainImage(ctx context.Context, request web.MerchantUpdateImageRequest, image interface{}) web.MerchantUpdateImageRequest {
+func (service *merchantServiceImpl) UpdateMainImage(ctx context.Context, request web.MerchantUpdateImageRequest) web.MerchantUpdateImageRequest {
 	merchant, err := service.MerchantRepository.FindById(ctx, request.Id)
 	helper.PanicIfError(err)
 
-	url, err := service.CloudinaryRepository.UploadImage(ctx, request.MainImage.FileName, image)
+	url, err := service.CloudinaryRepository.UploadImage(ctx, request.MainImage.FileName, request.MainImage.URL)
 	helper.PanicIfError(err)
 
 	_, err = service.MerchantRepository.Update(ctx, domain.Merchant{
 		Id:        merchant.Id,
 		UpdatedAt: request.UpdatedAt,
+		Slug:      merchant.Slug,
 		MainImage: &domain.Image{
 			Id:       merchant.MainImage.Id,
 			FileName: request.MainImage.FileName,
@@ -168,12 +182,20 @@ func (service *merchantServiceImpl) UpdateMainImage(ctx context.Context, request
 	err = service.CloudinaryRepository.DeleteImage(ctx, merchant.MainImage.FileName)
 	helper.PanicIfError(err)
 
+	request.MainImage.URL = url
 	return request
 }
 
 func (service *merchantServiceImpl) Delete(ctx context.Context, merchantId string) {
 	merchant, err := service.MerchantRepository.FindById(ctx, merchantId)
 	helper.PanicIfError(err)
+
+	products, err := service.ProductRepository.FindByMerchantId(ctx, merchant.Id.Hex())
+	helper.PanicIfError(err)
+
+	if products != nil {
+		panic("tidak dapat menghapus toko ini karena didalamnya masih terdapat produk")
+	}
 
 	err = service.MerchantRepository.Delete(ctx, merchant.Id.Hex())
 	helper.PanicIfError(err)
